# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * FinControl API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ApiError {
  error: string;
}

export interface SuccessMessage {
  message: string;
}

export interface AuthUser {
  id: number;
  username: string;
}

export type RegisterInputSecurityQuestion = typeof RegisterInputSecurityQuestion[keyof typeof RegisterInputSecurityQuestion];


export const RegisterInputSecurityQuestion = {
  pet: 'pet',
  escola: 'escola',
  cidade: 'cidade',
  mae: 'mae',
} as const;

export interface RegisterInput {
  /** @minLength 3 */
  username: string;
  /** @minLength 6 */
  password: string;
  securityQuestion: RegisterInputSecurityQuestion;
  /** @minLength 1 */
  securityAnswer: string;
}

export interface LoginInput {
  username: string;
  password: string;
}

export interface RecoverStep1Input {
  username: string;
}

export interface SecurityQuestion {
  question: string;
  questionKey: string;
}

export interface RecoverStep2Input {
  username: string;
  answer: string;
}

export interface TempPassword {
  tempPassword: string;
}

export type TransactionCategory = typeof TransactionCategory[keyof typeof TransactionCategory];


export const TransactionCategory = {
  venda: 'venda',
  material: 'material',
  funcionarios: 'funcionarios',
  outro: 'outro',
} as const;

export interface Transaction {
  id: number;
  /** YYYY-MM-DD */
  date: string;
  category: TransactionCategory;
  description: string;
  amount: number;
  createdAt: string;
}

export type TransactionInputCategory = typeof TransactionInputCategory[keyof typeof TransactionInputCategory];


export const TransactionInputCategory = {
  venda: 'venda',
  material: 'material',
  funcionarios: 'funcionarios',
  outro: 'outro',
} as const;

export interface TransactionInput {
  /** YYYY-MM-DD */
  date: string;
  category: TransactionInputCategory;
  description: string;
  /** @minimum 0.01 */
  amount: number;
}

export interface DashboardSummary {
  totalIncome: number;
  totalExpenses: number;
  balance: number;
  /** Percentage 0-100 */
  savingsRate: number;
  transactionCount: number;
  month: string;
}

export interface DayTotal {
  date: string;
  income: number;
  expenses: number;
}

export interface CategoryTotal {
  category: string;
  label: string;
  total: number;
  percentage: number;
}

export interface MonthlyReport {
  month: string;
  totalIncome: number;
  totalExpenses: number;
  balance: number;
  byCategory: CategoryTotal[];
  transactions: Transaction[];
}

export type ListTransactionsParams = {
category?: ListTransactionsCategory;
/**
 * YYYY-MM format
 */
month?: string;
limit?: number;
};

export type ListTransactionsCategory = typeof ListTransactionsCategory[keyof typeof ListTransactionsCategory];


export const ListTransactionsCategory = {
  venda: 'venda',
  material: 'material',
  funcionarios: 'funcionarios',
  outro: 'outro',
} as const;

export type GetMonthlyReportParams = {
/**
 * YYYY-MM format
 */
month: string;
};