# lib/api-zod/src/generated/types/dashboardSummary.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * FinControl API specification
 * OpenAPI spec version: 0.1.0
 */

export interface DashboardSummary {
  totalIncome: number;
  totalExpenses: number;
  balance: number;
  /** Percentage 0-100 */
  savingsRate: number;
  transactionCount: number;
  month: string;
}