# lib/api-zod/src/generated/types/dayTotal.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * FinControl API specification
 * OpenAPI spec version: 0.1.0
 */

export interface DayTotal {
  date: string;
  income: number;
  expenses: number;
}