# lib/api-zod/src/generated/types/monthlyReport.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * FinControl API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CategoryTotal } from './categoryTotal';
import type { Transaction } from './transaction';

export interface MonthlyReport {
  month: string;
  totalIncome: number;
  totalExpenses: number;
  balance: number;
  byCategory: CategoryTotal[];
  transactions: Transaction[];
}