# lib/api-zod/src/generated/types/recoverStep2Input.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * FinControl API specification
 * OpenAPI spec version: 0.1.0
 */

export interface RecoverStep2Input {
  username: string;
  answer: string;
}