# lib/api-zod/src/generated/types/registerInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * FinControl API specification
 * OpenAPI spec version: 0.1.0
 */
import type { RegisterInputSecurityQuestion } from './registerInputSecurityQuestion';

export interface RegisterInput {
  /** @minLength 3 */
  username: string;
  /** @minLength 6 */
  password: string;
  securityQuestion: RegisterInputSecurityQuestion;
  /** @minLength 1 */
  securityAnswer: string;
}