# lib/api-zod/src/generated/types/transaction.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * FinControl API specification
 * OpenAPI spec version: 0.1.0
 */
import type { TransactionCategory } from './transactionCategory';

export interface Transaction {
  id: number;
  /** YYYY-MM-DD */
  date: string;
  category: TransactionCategory;
  description: string;
  amount: number;
  createdAt: string;
}