# lib/api-zod/src/generated/types/transactionInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * FinControl API specification
 * OpenAPI spec version: 0.1.0
 */
import type { TransactionInputCategory } from './transactionInputCategory';

export interface TransactionInput {
  /** YYYY-MM-DD */
  date: string;
  category: TransactionInputCategory;
  description: string;
  /** @minimum 0.01 */
  amount: number;
}